fix : 중복 엔티티 병합
