--- backend/src/main/java/ku/hackerthon/BeMyMood/domain/member/Location.java
@@ -1,17 +1,0 @@
-package ku.hackerthon.BeMyMood.domain.member;
-
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Location {
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "location_id")
-    private Long id;
-
-    @Column(name = "location_name")
-    private String name;
-}
