+++ backend/src/main/java/ku/hackerthon/BeMyMood/domain/tag/Location.java
@@ -1,4 +1,4 @@
-package ku.hackerthon.BeMyMood.domain.member;
+package ku.hackerthon.BeMyMood.domain.tag;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
@@ -7,11 +7,11 @@ import javax.persistence.*;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Color {
+public class Location {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "color_id")
+    @Column(name = "location_id")
     private Long id;
 
-    @Column(name = "color_name")
+    @Column(name = "location_name")
     private String name;
 }
